+++ src/replication-protocol/upstream.ts
@@ -37,9 +37,24 @@ import {
  *   In contrast to the master, the fork can be assumed to never loose connection,
  *   so we do not have to prepare for missed out events.
  */
-export function startReplicationUpstream<RxDocType, CheckpointType>(
+export async function startReplicationUpstream<RxDocType, CheckpointType>(
     state: RxStorageInstanceReplicationState<RxDocType>
 ) {
+
+    if (
+        state.input.initialCheckpoint &&
+        state.input.initialCheckpoint.upstream
+    ) {
+        const checkpointDoc = await getLastCheckpointDoc(state, 'up');
+        if (!checkpointDoc) {
+            await setCheckpoint(
+                state,
+                'up',
+                state.input.initialCheckpoint.upstream
+            );
+        }
+    }
+
     const replicationHandler = state.input.replicationHandler;
     state.streamQueue.up = state.streamQueue.up.then(() => {
         return upstreamInitialSync().then(() => {